requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.15.2
protobuf==3.20
keras==2.2.4
opencv-python
h5py==2.10.0
imgaug
IPython[all]